+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/CompareLinkConverter.java
@@ -1,0 +1,10 @@
+package nbbrd.heylogs.spi;
+
+import lombok.NonNull;
+
+@FunctionalInterface
+public interface CompareLinkConverter {
+
+    @NonNull
+    CompareLink convert(@NonNull ProjectLink link) throws IllegalArgumentException;
+}