+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/ForgeLinkType.java
@@ -1,5 +1,5 @@
 package nbbrd.heylogs.spi;
 
 public enum ForgeLinkType {
-    COMMIT, COMPARE, ISSUE, REQUEST, MENTION, PROJECT
+    COMMIT, COMPARE, ISSUE, REQUEST, MENTION, REPOSITORY
 }