+++ heylogs-ext-forgejo/src/main/java/nbbrd/heylogs/ext/forgejo/Forgejo.java
@@ -27,6 +27,7 @@ public final class Forgejo implements Forge {
             .linkResolver(REQUEST, ForgejoRequestLink::resolve)
             .messageFetcher(ISSUE, ForgejoMessageFetcher.ISSUE)
             .messageFetcher(REQUEST, ForgejoMessageFetcher.REQUEST)
-            .linkParser(PROJECT, ForgejoRepositoryLink::parse)
+            .linkParser(REPOSITORY, ForgejoRepositoryLink::parse)
+            .compareLinkConverter(ForgejoCompareLink::of)
             .build();
 }