+++ heylogs-ext-github/src/main/java/nbbrd/heylogs/ext/github/GitHub.java
@@ -27,6 +27,7 @@ public final class GitHub implements Forge {
             .linkResolver(REQUEST, GitHubRequestLink::resolve)
             .messageFetcher(ISSUE, GitHubMessageFetcher.ISSUE)
             .messageFetcher(REQUEST, GitHubMessageFetcher.REQUEST)
-            .linkParser(PROJECT, GitHubRepositoryLink::parse)
+            .linkParser(REPOSITORY, GitHubRepositoryLink::parse)
+            .compareLinkConverter(GitHubCompareLink::of)
             .build();
 }