+++ heylogs-ext-github/src/main/java/nbbrd/heylogs/ext/github/GitHubRepositoryLink.java
@@ -27,7 +27,7 @@ class GitHubRepositoryLink implements GitHubProjectLink {
 
     @StaticFactoryMethod
     public static @NonNull GitHubRepositoryLink parse(@NonNull URL url) {
-        String[] pathArray = getPathArray(url);
+        String[] pathArray = getPathArray(url, true);
         checkPathLength(pathArray, 2);
         checkPathItem(pathArray, 0, OWNER_PATTERN);
         checkPathItem(pathArray, 1, REPO_PATTERN);