+++ heylogs-ext-github/src/test/java/nbbrd/heylogs/ext/github/GitHubRepositoryLinkTest.java
@@ -21,6 +21,11 @@ class GitHubRepositoryLinkTest {
         assertThat(link.getOwner()).isEqualTo("nbbrd");
         assertThat(link.getRepo()).isEqualTo("heylogs");
         assertThat(link.toURL().toString()).isEqualTo("https://github.com/nbbrd/heylogs");
+
+        GitHubRepositoryLink linkWithTrailingSlash = parse(urlOf("https://github.com/nbbrd/heylogs/"));
+        assertThat(linkWithTrailingSlash.getOwner()).isEqualTo("nbbrd");
+        assertThat(linkWithTrailingSlash.getRepo()).isEqualTo("heylogs");
+        assertThat(linkWithTrailingSlash.toURL().toString()).isEqualTo("https://github.com/nbbrd/heylogs");
     }
 
     @Test