+++ heylogs-ext-gitlab/src/main/java/nbbrd/heylogs/ext/gitlab/GitLab.java
@@ -27,6 +27,7 @@ public final class GitLab implements Forge {
             .linkResolver(REQUEST, GitLabRequestLink::resolve)
             .messageFetcher(ISSUE, GitLabMessageFetcher.ISSUE)
             .messageFetcher(REQUEST, GitLabMessageFetcher.REQUEST)
-            .linkParser(PROJECT, GitLabRepositoryLink::parse)
+            .linkParser(REPOSITORY, GitLabRepositoryLink::parse)
+            .compareLinkConverter(GitLabCompareLink::of)
             .build();
 }