+++ heylogs-ext-gitlab/src/main/java/nbbrd/heylogs/ext/gitlab/GitLabCompareLink.java
@@ -7,6 +7,7 @@ import nbbrd.design.RepresentableAs;
 import nbbrd.design.StaticFactoryMethod;
 import nbbrd.heylogs.spi.CompareLink;
 import nbbrd.heylogs.spi.ForgeRef;
+import nbbrd.heylogs.spi.ProjectLink;
 import org.jspecify.annotations.Nullable;
 
 import java.net.URL;
@@ -24,6 +25,16 @@ class GitLabCompareLink implements CompareLink, GitLabProjectLink {
     @StaticFactoryMethod
     public static @NonNull GitLabCompareLink parse(@NonNull URL url) {
         return parseLink(GitLabCompareLink::new, COMPARE_KEYWORD, THREE_DOT_DIFF_PATTERN, ThreeDotDiff::parse, url);
+    }
+
+    @StaticFactoryMethod
+    public static @NonNull GitLabCompareLink of(@NonNull ProjectLink link) {
+        if (link instanceof GitLabCompareLink) return (GitLabCompareLink) link;
+        if (link instanceof GitLabProjectLink) {
+            GitLabProjectLink gitLab = (GitLabProjectLink) link;
+            return new GitLabCompareLink(gitLab.getBase(), gitLab.getNamespace(), gitLab.getProject(), ThreeDotDiff.DEFAULT);
+        }
+        throw new IllegalArgumentException("Cannot create compare link from non-GitLab project link: " + link);
     }
 
     @NonNull